Offer help when selection spans nodes
This check ensures that we only ever do our naive regex replacement when it won't eff up the DOM. Say, for instance, that the page has this content, with the selection indicated by `^`:

     Studies imply that the anterior <a href="#">commissure</a> can be a compensatory pathway…
                            ^                                  ^

If we replace the selection with, say, "midline-crossing axon bundle", we'll end up with:

     Studies imply that the midline-crossing axon bundle can be a compensatory pathway…
                            ^                           ^

This breaks the link. There are circumstances where the current implementation (which just replaces every instance of "the" with a gradient-highlighted "the") deletes a little bit of a tag, but not the whole thing, resulting in a raw HTML fragment mid-page. This commit should go a long way to preventing such travesties thus making it easier, as well, to toggle the view state down the line.