🐞 Fixed: styles fail to install when running Svelte app
Solution: moved a blatant Law of Demeter violation from rdbl.js down into App.svelte, where it belongs. Basically, anything in rdbl.js isn't going to apply to the Svelte app by definition as that file is only included in manifest.json, so it's best to keep as much as possible under Svelte's hood where the code can be shared.