🐞 Fixed: hitting `tab` after entering Editor fails to send focus to the "definition" field
There's 2 things going on here.

* ARIA immprovements, as recommended [here](https://developer.mozilla.org/en-US/docs/Learn/Accessibility/HTML#Meaningful_text_labels).
* Use of `tabindex` + `autofucs` (plus Svelte's `ignore` tag to avoid getting a warning) to highlight the _label_.

My (limited) understanding is that assistive devices can provide pretty crappy experiences when focus jumps to an interactive element without first getting a chance to listen to some context. This might be "shortenable" so that the visual cue is ignored by assistive devices, but that there's some other cue leading them to hit the `/` trigger in the first place, in which case both experiences can be streamlined by a single keystroke…. This'll do for now. 🙄